Add files via upload
Problem 1: Write a python program that prints the following using print statement. 

Problem 2: You are given a grid as shown in the figure below. You can determine the color and number of each square from the grid. Write a python program that inputs two numbers (integer) within the grid range. Your program will determine if the two squares entered in this grid have the same color or not.  



Problem 3: The below mentioned figure shows a General diagnosis flowchart for troubleshooting HP  servers. Use the flowchart to create a python program that leads a person through the steps of fixing  a bad HP server. Here is an example of the program’s output (following one of the flow):

Starting General Diagnosis Program.

Recoding symptoms information - DONE.

Rebooting server to see if condition still exists - DONE.

Is this a newly installed server? yes [enter]

Please reseat any components that may have come loose during shipping - DONE. Rebooting the server - DONE.

Does the condition still exist? no [enter]

Recording all actions taken for future - DONE.

Congratulations, your server problems are solved.

National University of Computer & Emerging Sciences (NUCES),  

Islamabad Department of Computer Science

Notice the program ends as soon as a solution is found to the problem.



National University of Computer & Emerging Sciences (NUCES),  

Islamabad Department of Computer Science

Problem 4: Compose a python program to check if a year is leap year or not. If a year is divisible by  4 then it is leap year, however if the entered year is a century year like 2000, 1900, 2100 then it must  be distinct by 400.

Problem 5: Write a python program which accepts amount in rupees as input (integer) within Range  from Rs. 100 to Rs. 100000 and then display the total number of Currency Notes of Rs. 500, 100, 50,  20, 10, 5 and 1.

For example: when a user enters a number, Rs. 57477  

Currency Note : Number

500 : 114

100 : 4

50 : 1

20 : 1

10 : 0

5 : 1

1 : 2